requirements.txt
fire
numpy
pillow
scipy
tensorboard
torch
torchvision
tqdm
